pyproject.toml
[project]
name = "ingress2gateway"
version = "0.2.0"
description = "Convert Kubernetes Ingress objects to Gateway API resources"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "click>=8.1.0",
    "rich>=13.0.0",
    "jsonschema>=4.20.0",
]

[project.scripts]
ingress2gateway = "ingress2gateway.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line too long - HTML templates in strings

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E741"]  # Ambiguous variable name in tests
